requirements.txt
apache-airflow
apache-airflow-providers-apache-kafka
confluent-kafka